move opml parsing to separate module
